+++ openid-data-classes/src/commonMain/kotlin/at/asitplus/openid/dcql/DCQLCredential.kt
@@ -5,7 +5,7 @@ import at.asitplus.openid.CredentialFormatEnum
 sealed interface DCQLCredential {
     val claimStructure: DCQLCredentialClaimStructure
     val satisfiesCryptographicHolderBinding: Boolean
-    val authorityKeyIdentifiers: Collection<DCQLAuthorityKeyIdentifier>
     val format: CredentialFormatEnum
     val isSelectivelyDisclosable: Boolean
+    val authorityKeyIdentifiers: Collection<DCQLAuthorityKeyIdentifier>
 }