+++ CHANGELOG.md
@@ -22,6 +22,7 @@ Release 5.12.0 (unreleased):
    - Change: Don't send response on user initiated signature cancellation
    - BREAKING CHANGE: The result type from `verifyAuthnResponse`, `AuthnResponseResult` has been reworked to a data class
    - DCQL: Add custom credential types and proper satisfaction evaluation
+   - Add: DCQL submission requirements validation
  - OpenID for Verifiable Credential Issuance:
    - Moved the class `RefreshTokenInfo` from `OpenId4VciClient` to `SubjectCredentialStore.kt` and renamed it to `CredentialRenewalInfo` to better describe its role in the renewal process.
      Kept `RefreshTokenInfo` in the original package for backward compatibility
@@ -49,6 +50,10 @@ Release 5.12.0 (unreleased):
  - Dependencies:
    - Update to [Signum 3.21.0](https://github.com/a-sit-plus/signum/releases/tag/3.21.0) fixing CBOR parsing and tolerating cursed X.509 certificate encodings
    - Remove code elements deprecated in 5.11.0
+ - DCQL:
+   - Add: custom credential types
+   - Add: proper satisfaction evaluation
+   - Add: DCQL submission requirements validation
 
 Release 5.11.1:
  - OAuth 2.0: